--- src/utils/extractActionFromPath.js
@@ -1,5 +1,0 @@
-function extractActionFromPath(path) {
-  return path.split('/')[2];
-}
-
-export default extractActionFromPath;
