--- src/utils/extractQuoteIDFromPath.js
@@ -1,5 +1,0 @@
-function extractQuoteIDFromPath(path) {
-  return path.split('/')[3];
-}
-
-export default extractQuoteIDFromPath;
